requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
deepseek==1.0.0
distro==1.9.0
frozenlist==1.7.0
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
idna==3.10
jiter==0.10.0
magic-filter==1.0.12
multidict==6.6.4
openai==1.106.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
yarl==1.20.1
